requirement.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.115.0
filelock==3.19.1
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
huggingface_hub==1.1.4
idna==3.11
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.1.0
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tqdm==4.67.1
typer-slim==0.20.0
typing==3.7.4.3
typing_extensions==4.15.0
uvicorn==0.32.0
watchfiles==1.1.1
websockets==15.0.1
